basic_math: MOD: cast operands to 64-bit integers
No point in needlessly clamping the range.

+++ ttbasic/basic_math.cpp
@@ -200,7 +200,7 @@ Modulus operator.
       tmp = ivalue();
       if (err)
         return -1;
-      value = (int32_t)value % (int32_t)tmp; //割り算を実行
+      value = (int64_t)value % (int64_t)tmp; //割り算を実行
       break;
 
 /***bo op <<
